Uploads relationships

+++ src/infra/database/prisma/repositories/prisma-answer-attachments-repository.ts
@@ -1,28 +1,58 @@
 import { AnswerAttachmentsRepository } from '@/domain/forum/application/repositories/answer-attachments-repository';
 import { AnswerAttachment } from '@/domain/forum/enterprise/entities/answer-attachment';
 import { Injectable } from '@nestjs/common';
-import { PrismaAnswerAttachmenttMapper } from '../mappers/prisma-answer-attachment-mapper';
 import { PrismaService } from '../prisma.service';
+import { PrismaAnswerAttachmentMapper } from '../mappers/prisma-answer-attachment-mapper';
 
 @Injectable()
-export class PrismaAnswerAttachmentsRepository implements AnswerAttachmentsRepository {
-    constructor(private prisma:PrismaService) {}
+export class PrismaAnswerAttachmentsRepository
+implements AnswerAttachmentsRepository
+{
+    constructor(private prisma: PrismaService) {}
 
-    async findManyByAnswerId(questionId: string): Promise<AnswerAttachment[]> {
+    async findManyByAnswerId(answerId: string): Promise<AnswerAttachment[]> {
         const answerAttachments = await this.prisma.attachment.findMany({
             where: {
-                questionId
-            }
+                answerId,
+            },
         });
 
-        return answerAttachments.map(PrismaAnswerAttachmenttMapper.toDomain);
+        return answerAttachments.map(PrismaAnswerAttachmentMapper.toDomain);
     }
-    async deleteManyByAnswerId(questionId: string): Promise<void> {
+
+    async createMany(attachments: AnswerAttachment[]): Promise<void> {
+        if (attachments.length === 0) {
+            return;
+        }
+
+        const data = PrismaAnswerAttachmentMapper.toPrismaUpdateMany(attachments);
+
+        await this.prisma.attachment.updateMany(data);
+    }
+
+    async deleteMany(attachments: AnswerAttachment[]): Promise<void> {
+        if (attachments.length === 0) {
+            return;
+        }
+
+        const attachmentIds = attachments.map((attachment) => {
+            return attachment.id.toString();
+        });
+
         await this.prisma.attachment.deleteMany({
             where: {
-                questionId
-            }
+                id: {
+                    in: attachmentIds,
+                },
+            },
         });
     }
-    
+
+    async deleteManyByAnswerId(answerId: string): Promise<void> {
+        await this.prisma.attachment.deleteMany({
+            where: {
+                answerId,
+            },
+        });
+    }
 }
